my_resume/requirements.txt
channels==4.1.0
daphne==4.1.2
django-ckeditor==6.7.1
django-jsoneditor==0.2.4
markdown==3.6
pillow==10.3.0
python-dotenv==1.0.1
python-decouple==3.8
service-identity==24.1.0
zope.interface==6.3

google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0